Backend/requirements.txt
fastapi
langgraph
groq
streamlit
yfinance
newsapi-python
python-dotenv
uvicorn
reportlab
langchain-openai
requests